Move getTradingMode to separate file
