+++ packages/frontend/src/utils/instance.ts
@@ -1,5 +1,3 @@
-import { TradingMode } from "@explorer/shared"
-
 export type InstanceName = 'dYdX' | 'Myria' | 'GammaX'
 
 export function getInstanceName(): InstanceName {
@@ -14,20 +12,5 @@ export function getInstanceName(): InstanceName {
       return 'GammaX'
     default:
       throw new Error(`Unknown STARKEX_INSTANCE: ${projectName}`)
-  }
-}
-
-export function getTradingMode(): TradingMode {
-  const projectName = process.env.STARKEX_INSTANCE ?? 'dydx-mainnet'
-
-  switch (projectName) {
-    case 'dydx-mainnet':
-      return 'perpetual'
-    case 'myria-goerli':
-      return 'spot'
-    case 'gammax-goerli':
-      return 'perpetual'
-    default:
-      throw new Error(`Unknown TRADING_MODE: ${projectName}`)
   }
 }
